pyproject.toml
[project]
name = "faqsupportbot-backend"
version = "0.1.0"
description = "RAG-powered FAQ Support Bot Backend API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "langchain>=0.3.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=3.1.0",
    "langchain-qdrant>=1.1.0",
    "langchain-text-splitters>=1.0.0",
    "langchain-core>=0.3.0",
    "qdrant-client>=1.16.0",
    "python-dotenv>=1.0.0",
    "pypdf>=5.1.0",
    "unstructured[pdf]>=0.18.20",
    "python-multipart>=0.0.9",
    "streamlit>=1.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "black>=24.0.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"
